+++ src/test/java/io/supertokens/test/bulkimport/BulkImportTestUtils.java
@@ -41,6 +41,7 @@ import io.supertokens.thirdparty.InvalidProviderConfigException;
 import io.supertokens.totp.Totp;
 import io.supertokens.usermetadata.UserMetadata;
 import io.supertokens.userroles.UserRoles;
+import io.supertokens.utils.Utils;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -64,9 +65,8 @@ public class BulkImportTestUtils {
 
         for (int i = startIndex; i < numberOfUsers + startIndex; i++) {
             String email = "user" + i + "@example.com";
-            String id = io.supertokens.utils.Utils.getUUID();
-            String externalId = io.supertokens.utils.Utils.getUUID();
-
+            String externalId = io.supertokens.utils.Utils.getUUID() + "-" + i;
+            String id = "somebogus" + Utils.getUUID();
             JsonObject userMetadata = parser.parse("{\"key1\":\""+id+"\",\"key2\":{\"key3\":\"value3\"}}")
                     .getAsJsonObject();
 
@@ -81,13 +81,14 @@ public class BulkImportTestUtils {
             List<LoginMethod> loginMethods = new ArrayList<>();
             long currentTimeMillis = System.currentTimeMillis();
             loginMethods.add(new LoginMethod(tenants, "emailpassword", true, true, currentTimeMillis, email, "$2a",
-                    "BCRYPT", null, null, null, null));
+                    "BCRYPT", null, null, null, null, io.supertokens.utils.Utils.getUUID()));
             loginMethods
                     .add(new LoginMethod(tenants, "thirdparty", true, false, currentTimeMillis, email, null, null, null,
-                            "thirdPartyId" + i, "thirdPartyUserId" + i, null));
+                            "thirdPartyId" + i, "thirdPartyUserId" + i, null, io.supertokens.utils.Utils.getUUID()));
             loginMethods.add(
                     new LoginMethod(tenants, "passwordless", true, false, currentTimeMillis, email, null, null, null,
-                            null, null, null));
+                            null, null, null, io.supertokens.utils.Utils.getUUID()));
+            id = loginMethods.get(0).superTokensUserId;
             users.add(new BulkImportUser(id, externalId, userMetadata, userRoles, totpDevices, loginMethods));
         }
         return users;
@@ -116,13 +117,44 @@ public class BulkImportTestUtils {
             List<LoginMethod> loginMethods = new ArrayList<>();
             long currentTimeMillis = System.currentTimeMillis();
             loginMethods.add(new LoginMethod(tenants, "emailpassword", true, true, currentTimeMillis, email, null,
-                    null, "password"+i, null, null, null));
+                    null, "password"+i, null, null, null, io.supertokens.utils.Utils.getUUID()));
             loginMethods
                     .add(new LoginMethod(tenants, "thirdparty", true, false, currentTimeMillis, email, null, null, null,
-                            "thirdPartyId" + i, "thirdPartyUserId" + i, null));
+                            "thirdPartyId" + i, "thirdPartyUserId" + i, null, io.supertokens.utils.Utils.getUUID()));
             loginMethods.add(
                     new LoginMethod(tenants, "passwordless", true, false, currentTimeMillis, email, null, null, null,
-                            null, null, null));
+                            null, null, null, io.supertokens.utils.Utils.getUUID()));
+            id = loginMethods.get(0).superTokensUserId;
+            users.add(new BulkImportUser(id, externalId, userMetadata, userRoles, totpDevices, loginMethods));
+        }
+        return users;
+    }
+
+    public static List<BulkImportUser> generateBulkImportUserWithEmailPasswordAndRoles(int numberOfUsers, List<String> tenants, int startIndex, List<String> roles) {
+        List<BulkImportUser> users = new ArrayList<>();
+        JsonParser parser = new JsonParser();
+
+        for (int i = startIndex; i < numberOfUsers + startIndex; i++) {
+            String email = "user" + i + "@example.com";
+            String id = io.supertokens.utils.Utils.getUUID();
+            String externalId = io.supertokens.utils.Utils.getUUID();
+
+            JsonObject userMetadata = parser.parse("{\"key1\":\""+id+"\",\"key2\":{\"key3\":\"value3\"}}")
+                    .getAsJsonObject();
+
+            List<UserRole> userRoles = new ArrayList<>();
+            for(String roleName : roles) {
+                userRoles.add(new UserRole(roleName, tenants));
+            }
+
+            List<TotpDevice> totpDevices = new ArrayList<>();
+            totpDevices.add(new TotpDevice("secretKey", 30, 1, "deviceName"));
+
+            List<LoginMethod> loginMethods = new ArrayList<>();
+            long currentTimeMillis = System.currentTimeMillis();
+            loginMethods.add(new LoginMethod(tenants, "emailpassword", true, true, currentTimeMillis, email, null,
+                    null, "password"+i, null, null, null, io.supertokens.utils.Utils.getUUID()));
+            id = loginMethods.get(0).superTokensUserId;
             users.add(new BulkImportUser(id, externalId, userMetadata, userRoles, totpDevices, loginMethods));
         }
         return users;
@@ -168,6 +200,40 @@ public class BulkImportTestUtils {
         }
     }
 
+    public static void createTenantsWithinOneUserPool(Main main)
+            throws StorageQueryException, TenantOrAppNotFoundException, InvalidProviderConfigException,
+            FeatureNotEnabledException, IOException, InvalidConfigException,
+            CannotModifyBaseConfigException, BadPermissionException {
+        // User pool 1 - (null, null, null), (null, null, t1), (null, null, t2)
+
+        { // tenant 1
+            TenantIdentifier tenantIdentifier = new TenantIdentifier(null, null, "t1");
+
+            Multitenancy.addNewOrUpdateAppOrTenant(
+                    main,
+                    new TenantIdentifier(null, null, null),
+                    new TenantConfig(
+                            tenantIdentifier,
+                            new EmailPasswordConfig(true),
+                            new ThirdPartyConfig(true, null),
+                            new PasswordlessConfig(true),
+                            null, null, new JsonObject()));
+        }
+        { // tenant 2
+            TenantIdentifier tenantIdentifier = new TenantIdentifier(null, null, "t2");
+
+            Multitenancy.addNewOrUpdateAppOrTenant(
+                    main,
+                    new TenantIdentifier(null, null, null),
+                    new TenantConfig(
+                            tenantIdentifier,
+                            new EmailPasswordConfig(true),
+                            new ThirdPartyConfig(true, null),
+                            new PasswordlessConfig(true),
+                            null, null, new JsonObject()));
+        }
+    }
+
     public static void assertBulkImportUserAndAuthRecipeUserAreEqual(Main main, AppIdentifier appIdentifier,
             TenantIdentifier tenantIdentifier, Storage storage, BulkImportUser bulkImportUser,
             AuthRecipeUserInfo authRecipeUser) throws StorageQueryException, TenantOrAppNotFoundException {
@@ -179,6 +245,7 @@ public class BulkImportTestUtils {
             }
         }
         assertEquals(bulkImportUser.externalUserId, authRecipeUser.getSupertokensOrExternalUserId());
+        assertEquals(bulkImportUser.id, authRecipeUser.getSupertokensUserId());
         assertEquals(bulkImportUser.userMetadata,
                 UserMetadata.getUserMetadata(appIdentifier, storage, authRecipeUser.getSupertokensOrExternalUserId()));
 
@@ -195,6 +262,7 @@ public class BulkImportTestUtils {
     private static void assertLoginMethodEquals(Main main, io.supertokens.pluginInterface.authRecipe.LoginMethod lm1,
             io.supertokens.pluginInterface.bulkimport.BulkImportUser.LoginMethod lm2)
             throws TenantOrAppNotFoundException {
+        assertEquals(lm1.getSupertokensUserId(), lm2.superTokensUserId);
         assertEquals(lm1.email, lm2.email);
         assertEquals(lm1.verified, lm2.isVerified);
         assertTrue(lm2.tenantIds.containsAll(lm1.tenantIds) && lm1.tenantIds.containsAll(lm2.tenantIds));